email-scraper-service/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openai>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
